move all directory related moethds from RODSConn to RC_directory trait
